Test:	Use of MPU 6500 over i2c
This is a test script taken from:
http://blog.bitify.co.uk/2013/11/reading-data-from-mpu-6050-on-raspberry.html
and transposed into python3 (Adding brackets to `print` statements). Pieces of
this code will be used to form an update method for the buggy class
in buggy.py

Footer	I am using an MPU 6500 which, I know, can carry out Digital Motion Processing.
	....I need to learn how to use that data fully...